change inline to static inline when declare functions in /Csrc/*.h files
In \Csrc\*.h files, functions declared using inline, will be treated as undefined symbols when running the model (but no error when compiling). While change the 'inline' to 'static inline' or delete the 'inline' can fix this bug. This might be related with different versions of cython under python 2 and python 3.